+++ src/index.ts
@@ -1,5 +1,3 @@
-import { SupaliteClient } from './client';
-
 export * from './client';
 export * from './types';
 export * from './errors';